move existing REST server resources
