+++ dicoogle/src/main/java/pt/ua/dicoogle/server/ExtWebserviceApplication.java
@@ -16,9 +16,10 @@
  * You should have received a copy of the GNU General Public License
  * along with Dicoogle.  If not, see <http://www.gnu.org/licenses/>.
  */
-package pt.ua.dicoogle.webservices;
+package pt.ua.dicoogle.server;
 
 import java.util.ArrayList;
+import java.util.List;
 import org.slf4j.LoggerFactory;
 import org.restlet.Application;
 import org.restlet.Component;
@@ -27,6 +28,22 @@ import org.restlet.resource.ServerResource;
 import org.restlet.routing.Router;
 import pt.ua.dicoogle.server.web.rest.ExtResource;
 import pt.ua.dicoogle.server.web.rest.TestResource;
+import pt.ua.dicoogle.server.web.rest.ExamTimeResource;
+import pt.ua.dicoogle.server.web.rest.ExamTimeResource;
+import pt.ua.dicoogle.server.web.rest.ForceIndexing;
+import pt.ua.dicoogle.server.web.rest.ForceIndexing;
+import pt.ua.dicoogle.server.web.rest.RestDcmImageResource;
+import pt.ua.dicoogle.server.web.rest.RestDcmImageResource;
+import pt.ua.dicoogle.server.web.rest.RestDimResource;
+import pt.ua.dicoogle.server.web.rest.RestDimResource;
+import pt.ua.dicoogle.server.web.rest.RestDumpResource;
+import pt.ua.dicoogle.server.web.rest.RestDumpResource;
+import pt.ua.dicoogle.server.web.rest.RestFileResource;
+import pt.ua.dicoogle.server.web.rest.RestFileResource;
+import pt.ua.dicoogle.server.web.rest.RestTagsResource;
+import pt.ua.dicoogle.server.web.rest.RestTagsResource;
+import pt.ua.dicoogle.server.web.rest.RestWADOResource;
+import pt.ua.dicoogle.server.web.rest.RestWADOResource;
 
 /** A Restlet Application for aggregating extended services and 
  *
@@ -34,24 +51,28 @@ import pt.ua.dicoogle.server.web.rest.TestResource;
  * @author Luís A. Bastião Silva <bastiao@ua.pt>
  * @author Eduardo Pinho <eduardopinho@ua.pt>
  */
-public class WebservicePluginApplication extends Application {
+public class ExtWebserviceApplication extends Application {
 
-    static Component component = null;
-    static ArrayList<ServerResource> pluginServices = new ArrayList<>();
-
+    //static Component component = null;
+    private static final List<ServerResource> pluginServices = new ArrayList<>();
+    
     private Router internalRouter;
+    
+    public ExtWebserviceApplication() {
+        super();
+    }
 /*    
     public static void startWebservice(){
         try{
             
-            Logger.getLogger(WebservicePluginApplication.class.getName()).log(Level.INFO,
+            Logger.getLogger(ExtWebserviceApplication.class.getName()).log(Level.INFO,
                     "Adding Webservice in port {0}", ServerSettings.getInstance().getWeb().getServicePort());
             
             component = new Component();
             component.getServers().add(Protocol.HTTP,ServerSettings.getInstance().getWeb().getServicePort());
             
             // Attaches this application to the server.
-            component.getDefaultHost().attach(new WebservicePluginApplication());
+            component.getDefaultHost().attach(new ExtWebserviceApplication());
 
             // And starts the component.
             component.start();
@@ -84,7 +105,8 @@ public class WebservicePluginApplication extends Application {
         // new instance of our resources
         this.internalRouter = new Router(getContext());
         // Defines routing to resources
-        internalRouter.attach("/test/{something}{?query}", TestResource.class);
+        this.internalRouter.setDefaultMatchingQuery(false);
+        //internalRouter.attach("/test/{something}", TestResource.class);
         internalRouter.attach("/dim", RestDimResource.class);//search resource
         internalRouter.attach("/file", RestFileResource.class);//file download resource
         internalRouter.attach("/dump", RestDumpResource.class);//dump resource
@@ -103,7 +125,7 @@ public class WebservicePluginApplication extends Application {
         
         loadPlugins();
         
-        LoggerFactory.getLogger(WebservicePluginApplication.class).debug(internalRouter.getRoutes().toString());
+        LoggerFactory.getLogger(ExtWebserviceApplication.class).debug(internalRouter.getRoutes().toString());
         return internalRouter;
     }
     
@@ -111,7 +133,7 @@ public class WebservicePluginApplication extends Application {
         //lets add plugin registred services
         //this is still a little brittle... :(
         for(ServerResource resource : pluginServices) {
-            LoggerFactory.getLogger(WebservicePluginApplication.class).debug(
+            LoggerFactory.getLogger(ExtWebserviceApplication.class).debug(
                     "Inbound: {}", resource);
             internalRouter.attach("/" + resource.toString(), resource.getClass());
         }
