+++ dicoogle/src/main/java/pt/ua/dicoogle/server/web/rest/ExtResource.java
@@ -2,7 +2,7 @@ package pt.ua.dicoogle.server.web.rest;
 
 import org.restlet.data.Status;
 import org.restlet.representation.Representation;
-import org.restlet.representation.StringRepresentation;
+import org.restlet.resource.ResourceException;
 import org.restlet.resource.ServerResource;
 
 /**
@@ -12,7 +12,6 @@ import org.restlet.resource.ServerResource;
 public class ExtResource extends ServerResource {
     @Override
     protected Representation doHandle() {
-        this.setStatus(Status.CLIENT_ERROR_FORBIDDEN);
-        return new StringRepresentation("Extended services resource. Sub-resource interaction expected.");
+        throw new ResourceException(Status.CLIENT_ERROR_NOT_FOUND);
     }
 }
