+++ dicoogle/src/main/java/pt/ua/dicoogle/server/web/rest/RestCountQueryResults.java
@@ -16,7 +16,7 @@
  * You should have received a copy of the GNU General Public License
  * along with Dicoogle.  If not, see <http://www.gnu.org/licenses/>.
  */
-package pt.ua.dicoogle.webservices;
+package pt.ua.dicoogle.server.web.rest;
 
 import org.apache.commons.lang.NotImplementedException;
 import org.restlet.data.MediaType;
