+++ dicoogle/src/main/java/pt/ua/dicoogle/server/web/rest/RestImageResource.java
@@ -20,7 +20,7 @@
  * To change this template, choose Tools | Templates
  * and open the template in the editor.
  */
-package pt.ua.dicoogle.webservices;
+package pt.ua.dicoogle.server.web.rest;
 
 import java.io.*;
 import java.util.ArrayList;
