+++ dicoogle/src/main/java/pt/ua/dicoogle/server/web/rest/elements/ExamTimeCore.java
@@ -20,7 +20,7 @@
  * To change this template, choose Tools | Templates
  * and open the template in the editor.
  */
-package pt.ua.dicoogle.webservices.elements;
+package pt.ua.dicoogle.server.web.rest.elements;
 
 import java.io.File;
