+++ dicoogle/src/main/java/pt/ua/dicoogle/server/web/rest/elements/TExamTime.java
@@ -16,7 +16,7 @@
  * You should have received a copy of the GNU General Public License
  * along with Dicoogle.  If not, see <http://www.gnu.org/licenses/>.
  */
-package pt.ua.dicoogle.webservices.elements;
+package pt.ua.dicoogle.server.web.rest.elements;
 
 import java.io.BufferedWriter;
 import java.io.File;
